requirements.txt
python-dotenv
pandas
matplotlib
seaborn
